refactor: extract apply_patch engine/CLI into @bunny-agent/apply-patch
apps/daemon needed the standalone apply_patch CLI purely for its own
esbuild sidecar bundle, but the only way to get it was a devDependency
on @bunny-agent/runner-pi — even though the V4A parsing engine and CLI
have zero pi-specific code. Move them into a standalone package with
no pi dependency, and repoint runner-cli/daemon at it directly instead
of reaching through runner-pi.

No behavior change: runner-pi still wraps the shared engine as the
native apply_patch tool and the bash PATH shim; the Docker images and
esbuild build scripts are untouched.
