+++ apps/daemon/package.json
@@ -48,8 +48,8 @@
     "node": ">=20.0.0"
   },
   "devDependencies": {
+    "@bunny-agent/apply-patch": "workspace:*",
     "@bunny-agent/runner-harness": "workspace:*",
-    "@bunny-agent/runner-pi": "workspace:*",
     "@types/node": "^20.10.0",
     "esbuild": "^0.27.2",
     "fast-check": "^3.22.0",
