+++ packages/runner-pi/package.json
@@ -10,10 +10,6 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
-    },
-    "./apply-patch-bin": {
-      "types": "./dist/apply-patch-bin.d.ts",
-      "import": "./dist/apply-patch-bin.js"
     }
   },
   "files": [
@@ -27,6 +23,7 @@
     "test": "vitest run --passWithNoTests"
   },
   "dependencies": {
+    "@bunny-agent/apply-patch": "workspace:*",
     "@earendil-works/pi-agent-core": "0.80.7",
     "@earendil-works/pi-ai": "0.80.7",
     "@earendil-works/pi-coding-agent": "0.80.7",
